Update from https://github.com/Anime-Manga/references/commit/fc2a92746ebd8e41ce34e639d59c36e0c7ad56b0
