+++ src/references/Cesxhin.AnimeManga.Application/Interfaces/Services/IDescriptionBookService.cs
@@ -3,7 +3,7 @@ using Newtonsoft.Json.Linq;
 
 namespace Cesxhin.AnimeManga.Application.Interfaces.Services
 {
-    public interface IDescriptionBookService : IGeneralNameService<GenericMangaDTO, JObject>
+    public interface IDescriptionBookService : IGeneralNameService<GenericBookDTO, JObject>
     {
     }
 }
